Rewrite login with server component and cookies
